Add basic markdown rendering
Of necessity markdown rendering much be opinionated because Markdown is
badly specified (or more charitably, "deliberately left ambiguous") This
is great for users (no need to write strict syntax) but sucks for
implementors (often things are undefined or defined ambiguously). All of
which means it is practically impossible to parse Markdown and render it
back to an identical form without introducing a LOT of `text' nodes
capturing whitespace etc.

Anyway, this markdown rendrerer is adequate for now, but really both the
parser and renderer need to be revisited to allow more flexibility. The
current rendrerer makes a reasonable start on a 'pretty printer' though
;)